Add new delete-card-sku command
And for the kiler feature, add new can-i-build-this-deck command which checks a downloaded MTGO decklist against your collection and determines if you have the cards to build this deck and if not, the quantities of missing cards. Switches are available to exclude proxies and to only consider cards not already in any deck.