Drop our SplitView-based drawer implementation for dialogs based on DialogHost.Avalonia.
This mostly works except for a small quirk where opening a dialog while an existing open dialog is present requires us to "hide" that one and maintain a stack of open dialog ViewModels to restore the next open dialog when a dialog is closed. This is because I couldn't figure out how/if it is possible to show multiple dialogs at once.

This quirk aside, we are no longer bitten by the IsPaneOpen binding bug in SplitView that caused us absolute grief if we accidentally closed drawer content by clicking anywhere outside.

Fixes #21