Fix wishlist item changes not sticking around when immediately moved to a collection.
This is because we have multiple instances of DbContext living longer than they should be, meaning a saved wishlist item may not be appearing in other dbcontexts that may be querying for this same information.

To fix this, ensure all dbcontext accesses are true units of work with short-lived dbcontext instances. And the only way to do that is for our actual collection tracking service to DI in a Func<Owned<CardsDbContext>> and ensure dbcontext instances acquired through this are disposed of properly via C# using blocks.

Fixes #8